Moved IntegerIsIntegral to edu.gemini.shared.util package to avoid problems on export
